requirements.txt
spotivents @ https://www.github.com/justfoolingaround/spotivents.git
python-dotenv~=0.20.0
